+++ compiler/apps/playground/components/Editor/monacoOptions.ts
@@ -29,4 +29,5 @@ export const monacoOptions: Partial<EditorProps['options']> = {
   automaticLayout: true,
   wordWrap: 'on',
   wrappingIndent: 'same',
+  scrollBeyondLastLine: false,
 };
